refactor: Splitting plugin routes
Move over to using 3 subContainers, one for each possible kind of query, Packages/Titles/Joint package and title.

This commit handles the joint case ONLY

refs ERM-2916, ERM-2917, ERM-2918
