+++ src/Container/subContainers/PackageContainer.js
@@ -1,0 +1,6 @@
+/* This Container will handle the use case for Package plugin */
+const PackageContainer = () => {
+
+};
+
+export default PackageContainer;
