+++ src/Container/subContainers/TitleContainer.js
@@ -1,0 +1,6 @@
+/* This Container will handle the use case for Title plugin */
+const TitleContainer = () => {
+
+};
+
+export default TitleContainer;
