+++ src/Container/subContainers/index.js
@@ -1,0 +1,3 @@
+export { default as JointContainer } from './JointContainer';
+export { default as PackageContainer } from './PackageContainer';
+export { default as TitleContainer } from './TitleContainer';
